test-requirements.txt
--extra-index-url=https://wheelhouse.shopinvader.com/simple
# TODO locomotivecms dep should be in the setup.py of the addon that needs it
git+https://github.com/akretion/locomotivecms-python-client.git#egg=locomotivecms
# test requirements
requests_mock
vcrpy
vcrpy-unittest
openupgradelib @ git+https://github.com/OCA/openupgradelib.git
unittest2 # For shopinvader test_controller, which inherits component
# TMP dependency to include https://github.com/akretion/odoo-test-helper/pull/1
git+git://github.com/simahawk/odoo-test-helper.git@master#egg=odoo-test-helper
odoo-test-helper



# FIXME: TMP deps for shopinvader_sale_packaging
odoo13-addon-stock_packaging_calculator @ git+https://github.com/OCA/stock-logistics-warehouse@refs/pull/918/head#subdirectory=setup/stock_packaging_calculator
odoo13-addon-sale_order_line_packaging_qty @ git+https://github.com/OCA/sale-workflow@refs/pull/1114/head#subdirectory=setup/sale_order_line_packaging_qty
